Fixed datatype

+++ xilab/src/main/java/de/hrw/xilab/spring/model/api/LocationData.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class LocationData {
-    private long latitude;
-    private long longitude;
+    private Double latitude;
+    private Double longitude;
 
 }
